kernels: wahoo: apply a patch to fix the build
We need this patch after upstream stable commit 78d19b437554 ("kbuild:
fix single target build for external module") was merged via 4.4.210 in
the June security update. prepare is now called when building an
external module, which calls vdso_prepare, which then calls prepare0,
which is only defined when KBUILD_EXTMOD is set.

This is not a problem for upstream stable because commit
a66649dab350 ("arm64: fix vdso-offsets.h dependency") is not applied
there; it is only applied in the Wahoo tree as part of a series of vDSO
commits they added.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>